testbed/stacks/deserialization-python/app/requirements.txt
flask
pyyaml
jsonpickle
